tests: unify completely virtual test driver in _testutils
